requirements.txt
lxml
BeautifulSoup4
aiohttp
python-dateutil